refactor/fixing reservation Entity reration and create transaction

+++ src/modules/show/show.service.ts
@@ -1,39 +1,58 @@
-import { Injectable, UnauthorizedException, NotFoundException } from "@nestjs/common";
+import { 
+  Injectable, 
+  NotFoundException, 
+  ForbiddenException 
+} from "@nestjs/common";
 
 import { ShowRepository } from "src/modules/show/show.repository";
 
-import { ShowEntity } from "src/entities/shows.entity";
-import { UserEntity } from "src/entities/users.entity";
 import { CreateShowDto, UpdateShowDto } from "src/dto/show.dto";
-import { UserRole } from "src/common/constants/enums";
 
 @Injectable()
 export class ShowService {
   constructor(
-    private readonly showRepository: ShowRepository,
+    private readonly showRepository: ShowRepository
   ) {}
 
-  async createShow(createShowDto: CreateShowDto, user: UserEntity): Promise<ShowEntity> {
-    if (!user || user.role !== UserRole.HOST)
-      throw new UnauthorizedException("You do not have permission to create a show.");
-
-    const show = this.showRepository.createShow(createShowDto);
-    return show;
+  async createShow(createShowDto: CreateShowDto) {
+    return this.showRepository.createShow(createShowDto);
   }
 
-  async getShow(showId: number): Promise<ShowEntity[]> {
-    return this.showRepository.getShow(showId);
+  async getShowsAll() {
+    return this.showRepository.getShowsAll();
   }
 
-  async updateShow(updateShowDto: UpdateShowDto, user: UserEntity): Promise<ShowEntity> {
-    return this.showRepository.updateShow(updateShowDto);
+  async getShowsDetailsByGenre(genre: string) {
+    return this.showRepository.getShowsDetailsByGenre(genre);
   }
 
-  async deleteShow(showId: number): Promise<void> {
-   const show = await this.showRepository.deleteShow(showId);
+  async getShowDetails(showId: number) {
+    return this.showRepository.getShowDetails(showId);
+  }
 
-    if(show.affected === 0) {
-      throw new NotFoundException("Show not found");
-    }
+  async updateShow(
+    showId: number,
+    userId: number,
+    updateShowDto: UpdateShowDto,
+  ) {
+    const show = await this.showRepository.getShowDetails(showId);
+
+    // 공연 수정 권한 체크
+    if (show.userByUserEntity.userId !== userId)
+      throw new ForbiddenException("You are not allowed to update this show");
+
+    return this.showRepository.updateShow(updateShowDto, showId);
+  }
+
+  async deleteShow(showId: number, userId: number) {
+    const show = await this.showRepository.getShowDetails(showId);
+
+    // 공연 삭제 권한 체크
+    if (show.userByUserEntity.userId !== userId)
+      throw new ForbiddenException("You are not allowed to delete this show");
+
+    const deletedShow = await this.showRepository.deleteShow(showId);
+
+    if (deletedShow.affected === 0) throw new NotFoundException("Show not found");
   }
 }
